PT sep service: return the correct enc token length even in the failure case
In the API document, IHAGetOTP() is supposed to return the correct
length for the encrypted token even when IHAGetOTP() API call fails.
However the current implementation does not conform this API definition.
The fixes will require changes across sep service, libiha, and the
security firmware. This patch is the fix for sep service, and it needs to
be verified together with AOS patch #157466 for libiha, and
the IFWI with the fix for HSD #2342256.

Change-Id: Ia82b1052d70ea3b1e1d2b3e60b47df57eafd2ddb
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-458
Signed-off-by: Mark Allyn <mark.a.allyn@intel.com>
Reviewed-on: https://android.intel.com:443/246356